requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
cymem==2.0.5
Django==3.1.6
djangorestframework==3.12.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.2
MarkupSafe==1.1.1
murmurhash==1.0.5
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
parsel==1.6.0
pathy==0.4.0
preshed==3.0.5
Protego==0.1.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
smart-open==3.0.0
soupsieve==2.1
spacy==3.0.3
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==2.4.0
thinc==8.0.1
threadpoolctl==2.1.0
tqdm==4.56.2
Twisted==20.3.0
typer==0.3.2
urllib3==1.26.3
w3lib==1.22.0
wasabi==0.8.2
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.0.0/xx_ent_wiki_sm-3.0.0-py3-none-any.whl
zope.interface==5.2.0
